+++ LinearMetropolis/src/SimBox.h
@@ -46,9 +46,13 @@ public:
  	double makePeriodic(double x, double box);
  	double wrapBox(double x, double box);
  	void keepMoleculeInBox(Molecule *molecule, Environment *enviro);
+	//int * molecTables[];
+	Table * tables;
+	int molecTypenum;
+	int buildTable(int * tab);
  	int hopGE3(int atom1, int atom2, Molecule *molecule);
  	Molecule* getMoleculeFromAtomID(Atom *a1, Molecule *molecules, Environment *enviro);
- 	double getFValue(Atom *atom1, Atom *atom2, Molecule *molecules, Environment *enviro);
+ 	double getFValue(int atom1, int atom2, int **table1);//Atom *atom1, Atom *atom2, Molecule *molecules, Environment *enviro);
  	void generatefccBox(Molecule *molecules, Environment *enviro);
  	void generatePoints(Molecule *molecules, Environment *enviro);
  	void assignAtomPositions(double *dev_doublesX, double *dev_doublesY, double *dev_doublesZ, Molecule *molec, Environment *enviro);
