+++ LinearMetropolis/src/linearSim.h
