+++ parallelMetropolis/src/parallelSim.cu
