+++ parallelMetropolis/src/parallelUtil.cu
